moved benchmark into test_benchmark & start `--full-benchmark` option
